Add files via upload
the following code is for 3 methods analysis of the risk measurement parameters. We follow the same logic as ASX portfolio where we assume 3 methods the historical method, the parametric method and the montecarlo method.  a set of 5 portfolio stock is analysed : Apple, exon, Procter and Gamble, JP Morgan Chase, and Johnson and Johnson.